Add control of DRV8825 sleep mode to steppa.ino
The stepper motor gets hot at rest, unless the driver is put into sleep mode when not intended to be moving. It's possible this heating is a side effect of an incomplete step e.g. not leaving the step input in the correct state between steps, but better to avoid issue altogether in any case.